Modularise more, fix pages, reactivate term status
- Modularise some of the background script
- Modularise more of the content script, including the toolbar and
  stylesheet parts
  - Currently includes legacy hack for setting terms
- Make pages work again using modules
  - Highlighting in internal page not yet possible (if content script is
    injected and used, highlighting will not activate in any regular
    tabs)
- Make term occurrence checks work again, this time by putting the
  highlighting engine in charge of creating and invoking
  indicator/term-controls refresh generators
- Improve organisation of some code, mainly types
