+++ pages/options.html
@@ -4,11 +4,6 @@
 		<meta charset="utf-8">
 	</head>
 	<body>
-		<script src="/dist/include/utility.js"></script>
-		<script src="/dist/include/pattern-stem.js"></script>
-		<script src="/dist/include/pattern-diacritic.js"></script>
-		<script src="/dist/include/util-privileged.js"></script>
-		<script src="/dist/include/storage.js"></script>
-		<script src="/dist/pages/options.js"></script>
+		<script src="/dist/pages/options.mjs" type="module"></script>
 	</body>
 </html>
