+++ pages/startpage.html
@@ -4,14 +4,7 @@
 		<meta charset="utf-8">
 	</head>
 	<body>
-		<script src="/dist/include/utility.js"></script>
-		<script src="/dist/include/pattern-stem.js"></script>
-		<script src="/dist/include/pattern-diacritic.js"></script>
-		<script src="/dist/include/util-privileged.js"></script>
-		<script src="/dist/include/storage.js"></script>
-		<script src="/lib/email.min.js"></script>
-		<script src="/dist/include/page-build.js"></script>
-		<script src="/dist/pages/startpage-build.js"></script>
-		<script src="/dist/content.js"></script>
+		<script src="/dist/pages/startpage-build.mjs" type="module"></script>
+		<!--<script src="/dist/content.mjs" type="module"></script>-->
 	</body>
 </html>
