+++ src/modules/highlight/engines/paint/methods/paint.mts
@@ -1,4 +1,5 @@
-import type { Box, AbstractMethod } from "/dist/modules/highlight/engines/paint/method.mjs"
+import type { AbstractMethod } from "/dist/modules/highlight/engines/paint/method.mjs";
+import type { Box } from "/dist/modules/highlight/engines/paint.mjs";
 import type { AbstractHighlightability } from "/dist/modules/highlight/highlightability.mjs";
 import type { MatchTerm } from "/dist/modules/match-term.mjs";
 import { EleID, EleClass } from "/dist/modules/common.mjs";
@@ -91,4 +92,4 @@ class CSSPaintHighlightability implements AbstractHighlightability {
 	}
 }
 
-export { type TermSelectorStyles, PaintMethod }
+export { type TermSelectorStyles, PaintMethod };
