+++ src/modules/highlight/engines/paint/methods/url.mts
@@ -1,4 +1,5 @@
-import type { Box, AbstractMethod } from "/dist/modules/highlight/engines/paint/method.mjs"
+import type { AbstractMethod } from "/dist/modules/highlight/engines/paint/method.mjs";
+import type { Box } from "/dist/modules/highlight/engines/paint.mjs";
 import { StandardHighlightability } from "/dist/modules/highlight/highlightability.mjs";
 import type { MatchTerm } from "/dist/modules/match-term.mjs";
 import { EleID, EleClass } from "/dist/modules/common.mjs";
@@ -35,4 +36,4 @@ class UrlMethod implements AbstractMethod {
 	tempRemoveDrawElement = () => undefined;
 }
 
-export { UrlMethod }
+export { UrlMethod };
